refactor: rename speaker_verifier module to passphrase_verifier for naming consistency
